backend/requirements.txt
asgiref==3.2.7
certifi==2020.4.5.1
dj-database-url==0.5.0
Django==3.0.5
djangorestframework==3.11.0
gunicorn==20.0.4
psycopg2==2.8.5
Pygments==2.6.1
pytz==2019.3
sqlparse==0.3.1
whitenoise==5.0
